app/backend/requirements.txt
fastapi==0.111.0
email-validator==2.1.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
stripe==9.6.0
python-dotenv==1.0.1
httpx==0.27.0
requests==2.32.2
